+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fix quest log (fates) scroll up/down buttons.
+
 ## [v0.7.0] - 2021-08-19
 
 ### Fixed